requirements.txt
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.13.1
cycler==0.12.1
flake8==7.3.0
fonttools==4.61.1
imbalanced-learn==0.14.1
iniconfig==2.3.0
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
mccabe==0.7.0
nltk==3.9.2
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.3
